pyproject.toml
[tool.poetry]
name = "adsjust-test-assignments"
version = "0.1.0"
description = "This is a test assignment from the adjust company"
authors = ["Chashin Efim <69835249+NONVME@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.4"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
coverage = "^5.5"
pytest-cov = "^2.11.1"
yapf = "^0.30.0"
bpython = "^0.21"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
